work on SchoolCard and specific data to fetch and render
